requirements.txt
djangorestframework==3.15.1
requests==2.31.0
semver==3.0.2
